Separate Scene and Renderer.
Minor general include file organization
